ref: moved parseRawCustomers to customers directory
